requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
google-search-results>=2.4.2
anthropic>=0.25.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
